Starting to make the user api token controller
