+++ lib/siwapp_web/templates/user_api_token/new.html.heex
@@ -1,15 +1,6 @@
 <h1>API Token</h1>
 <p>Use the button below to generate an API token.</p>
-<.form let={f} for={@conn} action={Routes.user_api_token_path(@conn, :create)} as={:user}>
-  <%= label f, :email %>
-  <%= email_input f, :email, required: true %>
-
-  <%= label f, :password %>
-  <%= password_input f, :password, required: true %>
-
-  <%= label f, :remember_me, "Remember me on this computer" %>
-  <%= checkbox f, :remember_me %>
-
+<.form let={_f} for={@conn} action={Routes.user_api_token_path(@conn, :create)}>
   <div>
     <%= submit "Generate Token" %>
   </div>
